// src/webhooks/types.ts
export enum DeliveryMethod {
  Http = "http",
  EventBridge = "eventbridge",
}

type WebhookHandlerFunction = (
  topic: string,
  shop_domain: string,
  body: string
) => Promise<void>;

export interface RegisterOptions {
  // See https://shop0.dev/docs/admin-api/graphql/reference/events/webhooksubscriptiontopic for available topics
  topic: string;
  path: string;
  shop: string;
  accessToken: string;
  webhookHandler: WebhookHandlerFunction;
  deliveryMethod?: DeliveryMethod;
}

export interface RegisterReturn {
  success: boolean;
  result: unknown;
}

export interface WebhookRegistryEntry {
  path: string;
  topic: string;
  webhookHandler: WebhookHandlerFunction;
}

interface WebhookCheckResponseNode<
  T = {
    endpoint:
      | {
          __typename: "WebhookHttpEndpoint";
          callbackUrl: string;
        }
      | {
          __typename: "WebhookEventBridgeEndpoint";
          arn: string;
        };
  }
> {
  node: {
    id: string;
  } & T;
}

type WebhookCheckLegacyResponseNode = WebhookCheckResponseNode<{
  callbackUrl: string;
}>;

export interface WebhookCheckResponse<T = WebhookCheckResponseNode> {
  data: {
    webhookSubscriptions: {
      edges: T[];
    };
  };
}

export type WebhookCheckResponseLegacy = WebhookCheckResponse<
  WebhookCheckLegacyResponseNode
>;